Only write changed control point attributes
The common practice with RTPLAN DICOM files is to only write the control
point attributes that have changed since the previous ('current') value.
This makes for a more compact, but also easier to read DICOM file.
Perhaps this style is also more compatible, as it seems to be preferred
by the major TPS vendors.