Trying to get the scale situation right
Understanding how to deal with native readout format has been a big
pain.

First of all, it seems like its impossible to deduct from the RTP file
itself in all cases, how to correctly deal with scale conversion. I.e.
it seems to be impossible to know for sure which linac model is used,
and in some cases the scale conversion attribute is not even encoded in
the file because there are no control points. This is simply a big
weakness in Elekta's RTP format which causes such implementation
headaches.

To deal with this, instead of having the control point class trying to
figure out scale conversion, it is now left to the user to specify
either :varian or :elekta as a :scale option in the Plan#to_dcm method.

Although scale conversion is still possibly not understood too well,
probably the generation of Varian DICOM files are correct now. However,
for Elekta DICOM files, there may still be need for more investigation
before we can say that the implementation is correct.

The logic which checks for 'SYM' mode has been removed because it turned
out to be a dead end, with all invertion having to do with the
scale_convention attribute, and not the field mode.