Fix bug with misinterpretation of crc on old file
This fixes an issue that appeared when reading RTP files produced from
older Mosaiq versions than what is currently the latest verison
supported by the toolkit.

In these cases, the toolkit would mis-assign the crc value to the first
of the missing optional values in the parsed RTP record.

Thanks to Dennis for reporting this issue.