Added support for versioning in RTP output
Previously the rtp-connect toolkit would only write RTP files which
targetted the 'current' Mosaiq version. Obviously, this could lead to
problems if you wanted to use the latest version of the toolkit with an
older version of Mosaiq.

With this update, it is now possible to specify a :version option when
using the #write method. This update brings support for the following
Mosiaq versions: 2.4, 2.5 & 2.6.

Also fixed a bug where you couldn't access the ExtendedPlan record
belonging to a Plan record.