Set attributes dynamically for remainig two records
This implements dynamic attribute allocation to the control point and
dose tracking records, and reduces the total number of lines of code.

These two records still contain their own set_attribute methods though,
as opposed to the other record types where this method comes from the
Record class.

To simplify even further, perhaps it would be possible to merge this to
one set_attributes method for all records in the future.