Support for Mosaiq 2.64
This adds support for the new RTP file version apparently used in Mosaiq
2.64.

This version adds 3 attributes (iso_pos_x/y/z) in the Field and
ControlPoint records. Especially the ControlPoint record was somewhat
annoying, as the new values were not added at the end, as is the norm,
but rather in front of the MLC leaf position attributes.

This patch is based on an initial patch received from Stephan Ensminger.
Thank you for the contribution to the rtp-connect Ruby toolkit!

Note that as I dont have access to Mosaiq 2.64, this has not been
thouroughly tested with that enviornment yet.